Transfer gui_utils to the gui file
