--- shimmingtoolbox/gui_utils.py
@@ -1,25 +1,0 @@
-"""
-FSLeyes GUI utilities module.
-"""
-
-import imageio
-
-
-def imread(filename, bitdepth=8):
-    """Read image and convert it to desired bitdepth without truncation."""
-    if 'tif' in str(filename):
-        raw_img = imageio.imread(filename, format='tiff-pil')
-        if len(raw_img.shape) > 2:
-            raw_img = imageio.imread(filename, format='tiff-pil', as_gray=True)
-    else:
-        raw_img = imageio.imread(filename)
-        if len(raw_img.shape) > 2:
-            raw_img = imageio.imread(filename, as_gray=True)
-
-    img = imageio.core.image_as_uint(raw_img, bitdepth=bitdepth)
-    return img
-
-
-def imwrite(filename, img, format='png'):
-    """Write image."""
-    imageio.imwrite(filename, img, format=format)
